refactor: components removed due to shadcn config
